Update : ask for argument in "say" command

+++ bot.js
@@ -19,29 +19,47 @@ client.on("guildDelete", guild => {
 });
 
 client.on("message", async message => {
-    if(message.author.bot) return;
-    
-    if(message.content.indexOf(process.env.bot_prefix) !== 0) return;
+    if (message.author.bot) return;
+
+    if (message.content.indexOf(process.env.bot_prefix) !== 0) return;
 
     const args = message.content.slice(process.env.bot_prefix).trim().split(/ +/g);
     let command = args.shift().toLowerCase();
 
     command = command.replace(process.env.bot_prefix, '');
 
-    if(command === 'ping') {    
-      const m = await message.channel.send("Ping?");
-      m.edit(`Pong! Latency is ${m.createdTimestamp - message.createdTimestamp}ms. API Latency is ${Math.round(client.ws.ping)}ms`);
-    }
-    
-    if(command === "say") {
-      const sayMessage = args.join(" ");
-      message.delete().catch(O_o=>{});
-      message.channel.send(sayMessage);
+    if (command === 'ping') {
+        const m = await message.channel.send("Ping?");
+        m.edit(`Pong! Latency is ${m.createdTimestamp - message.createdTimestamp}ms. API Latency is ${Math.round(client.ws.ping)}ms`);
     }
 
-    if (command === 'cheatsheet'){
+    if (command === "say") {
+        const sayMessage = args.join(" ");
+        if (!sayMessage) {
+            //sayMessage = ``;
+            message.channel.send('What to say?');
+        }
+        else {
+            message.delete().catch(O_o => { });
+            message.channel.send(sayMessage);
+        }
+    }
+
+    /*
+        if (command === "shout") {
+            const sayMessage = args.join(" ");
+            if (!sayMessage) {
+                message.channel.send('Send a text to shout');
+            }
+            //message.delete().catch(O_o=>{});
+            else
+                message.channel.send(sayMessage);
+        }
+    */
+
+    if (command === 'cheatsheet') {
         // discord does not allow message longer than 2000 characters.
-        message.channel.send({files: ['cheatsheet.md']});
+        message.channel.send({ files: ['cheatsheet.md'] });
     }
 
 });
@@ -55,8 +73,8 @@ client.on('guildMemberAdd', member => {
     channel.send(message);
     // client.channels.cache.find('name', 'welcome').send(message);
 })
-  
-  
+
+
 let start = () => {
     client.login(process.env.bot_token);
 }
